requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
google-generativeai==0.4.1
python-dotenv==1.0.1
python-docx==1.1.0